Improve GraphQL error handling

+++ backend/src/API/GraphQL/Query/clickStreamQuery.js
@@ -1,119 +1,90 @@
 import { GraphQLString, GraphQLList } from "graphql";
 import {
-  authenticateAccessToken,
-  checkPermissionLevel,
   clickStreamCreator,
   getData,
   getDataFromDatabaseByFilter,
   nextID,
-  AllClickStreamsResponseModel,
-  CreateClickStreamModel,
-  NormalResponseModel,
 } from "../../../internal.js";
 import config from "../../../Config/config.js";
-
-function getResponseObject(message, statusCode, type) {
-  return {
-    message,
-    status: statusCode,
-    type,
-  };
-}
+import { assertIsAdmin, assertIsAuthenticated } from "../assert.js";
+import { UserInputError } from "../error.js";
+import { NormalResponseModel } from "../Model/model.js";
+import {
+  ClickStreamModel,
+  CreateClickStreamModel,
+} from "../Model/dataCollectionModel.js";
 
 // Root Queries - Used to retrieve data with GET-Requests
 const getAllClickStreamsQuery = {
-  type: AllClickStreamsResponseModel,
+  type: new GraphQLList(ClickStreamModel),
   args: {},
   async resolve(parent, args, context) {
-    try {
-      await authenticateAccessToken(context);
+    assertIsAuthenticated(context.user);
+    assertIsAdmin(context.user);
 
-      let response = checkPermissionLevel(
-        config.permissionLevel.admin,
-        context.user
+    const clickStreamsData = await getData(config.db.table.clickStream);
+    let clickStreams = [];
+    clickStreamsData.forEach((clickStreamData) => {
+      const clickStream = clickStreamCreator();
+      clickStreams.push(
+        clickStream.restoreObject(clickStream, clickStreamData)
       );
-      if (response.type === "error") {
-        return response;
-      }
+    });
 
-      const clickStreamsData = await getData(config.db.table.clickStream);
-      let clickStreams = [];
-      clickStreamsData.forEach((clickStreamData) => {
-        const clickStream = clickStreamCreator();
-        clickStreams.push(
-          clickStream.restoreObject(clickStream, clickStreamData)
-        );
-      });
-
-      response = getResponseObject(
-        "Click streams retrieved successfully",
-        200,
-        config.responseType.success
-      );
-
-      clickStreams = await Promise.all(clickStreams);
-      response.data = clickStreams;
-      return response;
-    } catch (error) {
-      return error;
-    }
+    clickStreams = await Promise.all(clickStreams);
+    return clickStreams;
   },
 };
 
 // Mutation Queries - Used to update or delete data with PUT- and DELETE-requests
 const createClickStreamQuery = {
-  type: NormalResponseModel,
+  type: ClickStreamModel,
   args: {
     sessionToken: { type: GraphQLString },
     clicks: { type: new GraphQLList(CreateClickStreamModel) },
   },
   async resolve(parent, args, context) {
-    try {
-      await authenticateAccessToken(context);
-      const { user } = context;
+    assertIsAuthenticated(context.user);
 
-      const refreshToken = await getDataFromDatabaseByFilter(
-        "id",
-        args.sessionToken,
-        config.db.table.refreshToken
-      );
+    const refreshToken = await getDataFromDatabaseByFilter(
+      "id",
+      args.sessionToken,
+      config.db.table.refreshToken
+    );
 
-      if (refreshToken == null) {
-        throw new Error("Session token is not valid.");
-      }
-
-      const clickStreamData = await getDataFromDatabaseByFilter(
-        "sessionToken",
-        args.sessionToken,
-        config.db.table.clickStream
-      );
-
-      let id;
-      let data;
-      const clickStream = clickStreamCreator();
-      if (clickStreamData == null) {
-        id = nextID(config.db.table.clickStream);
-        data = args;
-        data.id = id;
-        data.userID = user.id;
-        clickStream.updateData(data);
-      } else {
-        [data] = clickStreamData;
-        clickStream.updateData(data);
-        args.clicks.forEach((click) => {
-          clickStream.addClickNode(click);
-        });
-      }
-
-      return await clickStream.saveData(
-        "clickStream",
-        clickStream,
-        config.db.table.clickStream,
-        "Click stream stored successfully."
-      );
-    } catch (error) {
-      return error;
+    if (refreshToken == null) {
+      throw new Error("Session token is not valid.");
     }
+
+    const clickStreamData = await getDataFromDatabaseByFilter(
+      "sessionToken",
+      args.sessionToken,
+      config.db.table.clickStream
+    );
+
+    let id;
+    let data;
+    const clickStream = clickStreamCreator();
+    if (clickStreamData == null) {
+      id = nextID(config.db.table.clickStream);
+      data = args;
+      data.id = id;
+      data.userID = context.user.id;
+      clickStream.updateData(data);
+    } else {
+      [data] = clickStreamData;
+      clickStream.updateData(data);
+      args.clicks.forEach((click) => {
+        clickStream.addClickNode(click);
+      });
+    }
+    const response = await clickStream.saveData(
+      "clickStream",
+      clickStream,
+      config.db.table.clickStream,
+      "Click stream stored successfully."
+    );
+    return response;
   },
 };
 
@@ -123,37 +94,26 @@ const deleteClickStreamQuery = {
     clickStreamID: { type: GraphQLString },
   },
   async resolve(parent, args, context) {
-    try {
-      await authenticateAccessToken(context);
-      const { user } = context;
+    assertIsAuthenticated(context.user);
+    assertIsAdmin(context.user);
 
-      let response = checkPermissionLevel(config.permissionLevel.admin, user);
-      if (response.type === "error") {
-        return response;
-      }
+    const clickStreamData = await getDataFromDatabaseByFilter(
+      "id",
+      args.clickStreamID,
+      config.db.table.clickStream
+    );
 
-      const clickStreamData = await getDataFromDatabaseByFilter(
-        "id",
-        args.clickStreamID,
-        config.db.table.clickStream
+    if (clickStreamData == null || clickStreamData[0] == null) {
+      throw new UserInputError(
+        "The click stream you try to delete does not exist."
       );
-
-      if (clickStreamData == null || clickStreamData[0] == null) {
-        return getResponseObject(
-          "The click stream you try to delete does not exist.",
-          400,
-          config.responseType.error
-        );
-      }
-
-      const clickStream = clickStreamCreator();
-      clickStream.updateData(clickStreamData[0]);
-
-      response = await clickStream.deleteClickStream();
-      return response;
-    } catch (error) {
-      return error;
     }
+
+    const clickStream = clickStreamCreator();
+    clickStream.updateData(clickStreamData[0]);
+
+    const message = await clickStream.deleteClickStream();
+    return { message };
   },
 };
